requirements.txt
appnope==0.1.0
attrs==19.1.0
autopep8==1.5
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
idna==2.8
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.14.1
Jinja2==2.10.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
lxml==4.5.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.0
numpy==1.18.1
pandas==1.0.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.3
pyzmq==18.1.0
qtconsole==4.5.2
requests==2.22.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9.2
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.2
urllib3==1.25.8
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
